--- src/main/java/com/mobile/domain/auth/dto/SignInReq.java
@@ -1,15 +1,0 @@
-package com.mobile.domain.auth.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SignInReq {
-    @NotBlank
-    String studentId;
-
-    @NotBlank
-    String password;
-}
