--- src/main/java/com/mobile/domain/auth/dto/UserDto.java
@@ -1,12 +1,0 @@
-package com.mobile.domain.auth.dto;
-
-import com.mobile.domain.auth.entity.RoleType;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserDto {
-    private Long id;
-    private RoleType role;
-}
