--- src/main/java/com/mobile/domain/auth/entity/RoleType.java
@@ -1,12 +1,0 @@
-package com.mobile.domain.auth.entity;
-
-import org.springframework.security.core.GrantedAuthority;
-
-public enum RoleType implements GrantedAuthority {
-    STUDENT, ADMIN;
-
-    @Override
-    public String getAuthority() {
-        return "ROLE_" + this.name();
-    }
-}
