--- src/main/java/com/mobile/domain/auth/jwt/TokenProvider.java
@@ -1,8 +1,0 @@
-package com.mobile.domain.auth.jwt;
-
-public interface TokenProvider {
-    String createAccessToken(Long userId, String role);
-    String createRefreshToken(Long userId, String rotateId);
-    Long parseUserId(String token);
-    String parseRole(String token);
-}
