--- src/main/java/com/mobile/domain/auth/jwt/TokenType.java
@@ -1,6 +1,0 @@
-package com.mobile.domain.auth.jwt;
-
-public enum TokenType {
-    ACCESS,
-    REFRESH
-}
