--- src/main/java/com/mobile/domain/auth/refresh/RefreshCookie.java
@@ -1,4 +1,0 @@
-package com.mobile.domain.auth.refresh;
-
-public class RefreshCookie {
-}
