--- src/main/java/com/mobile/domain/auth/refresh/RefreshTokenService.java
@@ -1,4 +1,0 @@
-package com.mobile.domain.auth.refresh;
-
-public class RefreshTokenService {
-}
