--- src/main/java/com/mobile/domain/auth/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.mobile.domain.auth.repository;
-
-import com.mobile.domain.auth.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByStudentId(String studentId);
-    boolean existsByStudentId(String studentId);
-}
