+++ src/main/java/com/mobile/server/config/SwaggerConfig.java
@@ -1,0 +1,21 @@
+package com.mobile.server.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+
+
+    @Bean
+    public OpenAPI customOpenAPI() {
+        return new OpenAPI()
+                .info(new Info()
+                        .title("Mobile Server API")
+                        .description("Restful API documentation for Mobile Server")
+                        .version("1.0")
+                );
+    }
+}
