+++ src/main/java/com/mobile/server/domain/auth/dto/SignInReq.java
@@ -1,0 +1,15 @@
+package com.mobile.server.domain.auth.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class SignInReq {
+    @NotBlank
+    String studentId;
+
+    @NotBlank
+    String password;
+}
