+++ src/main/java/com/mobile/server/domain/auth/dto/SignUpReq.java
@@ -1,0 +1,24 @@
+package com.mobile.server.domain.auth.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class SignUpReq {
+    @NotBlank
+    @Size(min = 2, max = 30)
+    String nickname;
+
+    @NotBlank
+    String studentId;
+
+    @NotBlank
+    String password;
+}
