+++ src/main/java/com/mobile/server/domain/auth/dto/UserDto.java
@@ -1,0 +1,12 @@
+package com.mobile.server.domain.auth.dto;
+
+import com.mobile.server.domain.auth.entity.RoleType;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class UserDto {
+    private Long id;
+    private RoleType role;
+}
