+++ src/main/java/com/mobile/server/domain/auth/jwt/TokenProvider.java
@@ -1,0 +1,8 @@
+package com.mobile.server.domain.auth.jwt;
+
+public interface TokenProvider {
+    String createAccessToken(Long userId, String role);
+    String createRefreshToken(Long userId, String rotateId);
+    Long parseUserId(String token);
+    String parseRole(String token);
+}
