+++ src/main/java/com/mobile/server/domain/auth/jwt/TokenType.java
@@ -1,0 +1,6 @@
+package com.mobile.server.domain.auth.jwt;
+
+public enum TokenType {
+    ACCESS,
+    REFRESH
+}
