+++ src/main/java/com/mobile/server/domain/auth/refresh/RefreshToken.java
@@ -1,0 +1,4 @@
+package com.mobile.server.domain.auth.refresh;
+
+public class RefreshToken {
+}
