+++ src/main/java/com/mobile/server/domain/auth/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.mobile.server.domain.auth.repository;
+
+import com.mobile.server.domain.auth.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByStudentId(String studentId);
+    boolean existsByStudentId(String studentId);
+}
