+++ src/main/java/com/mobile/server/domain/auth/service/UserService.java
@@ -1,0 +1,46 @@
+package com.mobile.server.domain.auth.service;
+
+import com.mobile.server.domain.auth.dto.SignUpReq;
+import com.mobile.server.domain.auth.entity.RoleType;
+import com.mobile.server.domain.auth.entity.User;
+import com.mobile.server.domain.auth.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class UserService {
+
+    private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
+
+    //회원가입
+    @Transactional
+    public void register(SignUpReq signUpReq) {
+        if (userRepository.existsByStudentId(signUpReq.getStudentId())) {
+            throw new IllegalArgumentException("이미 존재하는 아이디입니다.");
+        }
+
+        User user = User.builder()
+                .studentId(signUpReq.getStudentId())
+                .password(passwordEncoder.encode(signUpReq.getPassword()))
+                .nickname(signUpReq.getNickname())
+                .role(RoleType.STUDENT)
+                .build();
+
+        userRepository.save(user);
+    }
+
+    //로그인 사용자 검증
+    @Transactional(readOnly = true)
+    public User verify(String studentId, String rawPassword) {
+        User user = userRepository.findByStudentId(studentId)
+                .orElseThrow(() -> new IllegalArgumentException("아이디 또는 비밀번호가 올바르지 않습니다."));
+        if (!passwordEncoder.matches(rawPassword, user.getPassword())) {
+            throw new IllegalArgumentException("아이디 또는 비밀번호가 올바르지 않습니다.");
+        }
+        return user;
+    }
+}
