+++ src/main/resources/META-INF/spring.factories
@@ -1,1 +1,1 @@
-org.springframework.boot.env.EnvironmentPostProcessor=com.mobile.common.config.DotenvEnvironmentPostProcessor
+org.springframework.boot.env.EnvironmentPostProcessor=com.mobile.server.config.DotenvEnvironmentPostProcessor
